requirements.txt
web3==6.10.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
setuptools